requirements.txt
streamlit
pandas
numpy
pillow
tensorflow
requests
lxml